Restructure command results
