+++ projects/01-rubygems-cli/solution/lib/command_results/command_result.rb
