+++ projects/01-rubygems-cli/solution/lib/search_command.rb
@@ -1,25 +1,7 @@
-require './lib/command_result'
+require './lib/command_results/search_command_result'
 require './lib/ruby_gems_api'
 
 class SearchCommand
-  class SearchCommandResult < CommandResult
-    attr_accessor :gem_names
-
-    def initialize(gem_names, exit_code = 0)
-      super(exit_code)
-      @gem_names = gem_names
-    end
-
-    def output
-      result = ''
-      gem_names.each_with_index do |name, i|
-        result += "#{i + 1}. #{name}\n"
-      end
-      
-      result
-    end
-  end
-
   class << self
     def execute(arg)
       result = RubyGemsApi.search_gems(arg)
