+++ projects/01-rubygems-cli/solution/lib/show_command.rb
@@ -1,21 +1,7 @@
-require './lib/command_result'
+require './lib/command_results/show_command_result'
 require './lib/ruby_gems_api'
 
 class ShowCommand
-  class ShowCommandResult < CommandResult
-    attr_accessor :gem_name, :info
-
-    def initialize(gem_name, info, exit_code = 0)
-      super(exit_code)
-      @gem_name = gem_name
-      @info = info
-    end
-
-    def output
-      "#{gem_name}\n#{info}"
-    end
-  end
-
   class << self
     def execute(arg)
       result = RubyGemsApi.gem_info(arg)
